requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.0
opencv-python==4.9.0.80
pillow==10.2.0
ultralytics==8.1.0
roboflow==1.1.23
pyyaml==6.0.1
torch>=2.0.0
torchvision>=0.15.0
